feat: 35 Application Structure - Rename packages
Main package name changed from com.event.scraper
to hub.event.
POM groupId and artifactId also modified accordingly.
