+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>2.7.1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.event</groupId>
-	<artifactId>scraper</artifactId>
+	<groupId>hub.event</groupId>
+	<artifactId>event-hub</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>scrapper</name>
-	<description>Cultural events scraper</description>
+	<description>Cultural events hub</description>
 	<properties>
 		<java.version>17</java.version>
 	</properties>
