--- src/main/java/com/event/scraper/ScraperApplication.java
@@ -1,13 +1,0 @@
-package com.event.scraper;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ScraperApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ScraperApplication.class, args);
-	}
-
-}
