--- src/main/java/com/event/scraper/adminapi/AdminController.java
@@ -1,10 +1,0 @@
-package com.event.scraper.adminapi;
-
-import com.event.scraper.auth.AuthService;
-import com.event.scraper.users.UserService;
-import com.event.scraper.statistics.StatsService;
-
-class AdminController {
-
-
-}
