--- src/main/java/com/event/scraper/auth/AuthService.java
@@ -1,4 +1,0 @@
-package com.event.scraper.auth;
-
-public class AuthService {
-}
