--- src/main/java/com/event/scraper/events/EventService.java
@@ -1,3 +1,0 @@
-package com.event.scraper.events;
- public class EventService {
-}
