--- src/main/java/com/event/scraper/mailmodule/MailService.java
@@ -1,4 +1,0 @@
-package com.event.scraper.mailmodule;
-
-public class MailService {
-}
