--- src/main/java/com/event/scraper/newsletter/NewsLetterService.java
@@ -1,7 +1,0 @@
-package com.event.scraper.newsletter;
-
-import com.event.scraper.mailmodule.MailService;
-import com.event.scraper.events.EventService;
-
-public class NewsLetterService {
-}
