--- src/main/java/com/event/scraper/scrapers/ScrapersService.java
@@ -1,8 +1,0 @@
-package com.event.scraper.scrapers;
-
-import com.event.scraper.events.EventService;
-
-
-public class ScrapersService {
-
-}
