--- src/main/java/com/event/scraper/scrapers/Scrappable.java
@@ -1,4 +1,0 @@
-package com.event.scraper.scrapers;
-
-public interface Scrappable {
-}
