--- src/main/java/com/event/scraper/scrapers/ebilet/EbiletScraper.java
@@ -1,6 +1,0 @@
-package com.event.scraper.scrapers.ebilet;
-
-import com.event.scraper.scrapers.Scrappable;
-
-class EbiletScraper implements Scrappable {
-}
