--- src/main/java/com/event/scraper/scrapers/empikbilet/EmpikBiletScraper.java
@@ -1,6 +1,0 @@
-package com.event.scraper.scrapers.empikbilet;
-
-import com.event.scraper.scrapers.Scrappable;
-
-class EmpikBiletScraper implements Scrappable {
-}
