--- src/main/java/com/event/scraper/scrapers/goingapp/GoingAppScraper.java
@@ -1,7 +1,0 @@
-package com.event.scraper.scrapers.goingapp;
-
-import com.event.scraper.scrapers.Scrappable;
-
-
-class GoingAppScraper implements Scrappable {
-}
