--- src/main/java/com/event/scraper/scrapers/kupbilecik/KupBilecikScraper.java
@@ -1,9 +1,0 @@
-package com.event.scraper.scrapers.kupbilecik;
-
-import com.event.scraper.scrapers.Scrappable;
-
-
-class KupBilecikScraper implements Scrappable {
-
-
-}
