--- src/main/java/com/event/scraper/scrapers/proanima/ProanimaScraper.java
@@ -1,6 +1,0 @@
-package com.event.scraper.scrapers.proanima;
-
-import com.event.scraper.scrapers.Scrappable;
-
-class ProanimaScraper implements Scrappable {
-}
