--- src/main/java/com/event/scraper/statistics/StatsService.java
@@ -1,8 +1,0 @@
-package com.event.scraper.statistics;
-
-import com.event.scraper.events.EventService;
-import com.event.scraper.scrapers.ScrapersService;
-import com.event.scraper.users.UserService;
-
-public class StatsService {
-}
