--- src/main/java/com/event/scraper/users/UserService.java
@@ -1,8 +1,0 @@
-package com.event.scraper.users;
-
-import com.event.scraper.auth.AuthService;
-
-public class UserService {
-
-
-}
