--- src/main/java/com/event/scraper/usersapi/UserController.java
@@ -1,8 +1,0 @@
-package com.event.scraper.usersapi;
-
-import com.event.scraper.auth.AuthService;
-import com.event.scraper.users.UserService;
-import com.event.scraper.newsletter.NewsLetterService;
-import com.event.scraper.events.EventService;
-class UserController {
-}
