+++ src/main/java/hub/event/ScraperApplication.java
@@ -1,0 +1,13 @@
+package hub.event;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ScraperApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ScraperApplication.class, args);
+	}
+
+}
