+++ src/main/java/hub/event/adminapi/AdminController.java
@@ -1,0 +1,10 @@
+package hub.event.adminapi;
+
+import hub.event.auth.AuthService;
+import hub.event.users.UserService;
+import hub.event.statistics.StatsService;
+
+class AdminController {
+
+
+}
