+++ src/main/java/hub/event/auth/AuthService.java
@@ -1,0 +1,4 @@
+package hub.event.auth;
+
+public class AuthService {
+}
