+++ src/main/java/hub/event/events/EventService.java
@@ -1,0 +1,3 @@
+package hub.event.events;
+ public class EventService {
+}
