+++ src/main/java/hub/event/mailmodule/MailService.java
@@ -1,0 +1,4 @@
+package hub.event.mailmodule;
+
+public class MailService {
+}
