+++ src/main/java/hub/event/newsletter/NewsLetterService.java
@@ -1,0 +1,7 @@
+package hub.event.newsletter;
+
+import hub.event.mailmodule.MailService;
+import hub.event.events.EventService;
+
+public class NewsLetterService {
+}
