+++ src/main/java/hub/event/scrapers/ScrapersService.java
@@ -1,0 +1,7 @@
+package hub.event.scrapers;
+
+import hub.event.events.EventService;
+
+public class ScrapersService {
+
+}
