+++ src/main/java/hub/event/scrapers/Scrappable.java
@@ -1,0 +1,4 @@
+package hub.event.scrapers;
+
+public interface Scrappable {
+}
