+++ src/main/java/hub/event/scrapers/ebilet/EbiletScraper.java
@@ -1,0 +1,6 @@
+package hub.event.scrapers.ebilet;
+
+import hub.event.scrapers.Scrappable;
+
+class EbiletScraper implements Scrappable {
+}
