+++ src/main/java/hub/event/scrapers/empikbilet/EmpikBiletScraper.java
@@ -1,0 +1,6 @@
+package hub.event.scrapers.empikbilet;
+
+import hub.event.scrapers.Scrappable;
+
+class EmpikBiletScraper implements Scrappable {
+}
