+++ src/main/java/hub/event/scrapers/goingapp/GoingAppScraper.java
@@ -1,0 +1,7 @@
+package hub.event.scrapers.goingapp;
+
+import hub.event.scrapers.Scrappable;
+
+
+class GoingAppScraper implements Scrappable {
+}
