+++ src/main/java/hub/event/scrapers/kupbilecik/KupBilecikScraper.java
@@ -1,0 +1,9 @@
+package hub.event.scrapers.kupbilecik;
+
+import hub.event.scrapers.Scrappable;
+
+
+class KupBilecikScraper implements Scrappable {
+
+
+}
