+++ src/main/java/hub/event/scrapers/proanima/ProanimaScraper.java
@@ -1,0 +1,6 @@
+package hub.event.scrapers.proanima;
+
+import hub.event.scrapers.Scrappable;
+
+class ProanimaScraper implements Scrappable {
+}
