+++ src/main/java/hub/event/statistics/StatsService.java
@@ -1,0 +1,8 @@
+package hub.event.statistics;
+
+import hub.event.events.EventService;
+import hub.event.scrapers.ScrapersService;
+import hub.event.users.UserService;
+
+public class StatsService {
+}
