+++ src/main/java/hub/event/users/UserService.java
@@ -1,0 +1,6 @@
+package hub.event.users;
+
+import hub.event.auth.AuthService;
+
+public class UserService {
+}
