+++ src/main/java/hub/event/usersapi/UserController.java
@@ -1,0 +1,9 @@
+package hub.event.usersapi;
+
+import hub.event.auth.AuthService;
+import hub.event.users.UserService;
+import hub.event.newsletter.NewsLetterService;
+import hub.event.events.EventService;
+
+class UserController {
+}
