--- src/test/java/com/event/scraper/ScraperApplicationTests.java
@@ -1,13 +1,0 @@
-package com.event.scraper;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ScraperApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
