+++ src/test/java/hub/event/ScraperApplicationTests.java
@@ -1,0 +1,13 @@
+package hub.event;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ScraperApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
